Alpha

+++ Alpha/1TypeCasting/TypePromoExampl.java
@@ -3,7 +3,8 @@ public class TypePromoExampl {
         char a = 'b';
         byte b = 1;
         short c = 2;
-        short d = a + b + c; // char is promoted to int so gives error: incompatible types: possible lossy conversion from int to short
+        //short d = a + b + c; // char is promoted to int so gives error: incompatible types: possible lossy conversion from int to short
+        short d = (short) (a + b + c);
         System.out.println(d);
     }
 }
